requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cloudinary==1.30.0
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
